Add dedicated Inventory page with sheet history and label search
Moves Sheet History and Label Search off the upload page onto a new
/inventory route, keeping the upload page focused on label generation.
