+++ frontend-vite/react-ts/src/App.tsx
@@ -7,6 +7,7 @@ import ProtectedRoute from './config/ProtectedRoute';
 import Help from './components/help';
 import SignUp from './components/signup';
 import LandingPage from './pages/LandingPage';
+import Inventory from './components/Inventory';
 
 function UploadPage() {
   return (
@@ -14,6 +15,17 @@ function UploadPage() {
       <Sidebar />
       <main className="flex-1 min-w-0">
         <LabelUploader />
+      </main>
+    </div>
+  );
+}
+
+function InventoryPage() {
+  return (
+    <div className="flex flex-col md:flex-row min-h-screen bg-background">
+      <Sidebar />
+      <main className="flex-1 min-w-0">
+        <Inventory />
       </main>
     </div>
   );
@@ -38,6 +50,14 @@ function App() {
           element={
             <ProtectedRoute>
               <UploadPage />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/inventory"
+          element={
+            <ProtectedRoute>
+              <InventoryPage />
             </ProtectedRoute>
           }
         />
